Add files via upload
Added serializer file. The serializers in REST framework work very similarly to Django's Form and ModelForm classes. Serializer gives you a powerful, generic way to control the output of your responses, as well as a ModelSerializer class which provides a useful shortcut for creating serializers that deal with model instances and query sets.